api/requirements.txt
pytest
pylint
requests
psycopg2-binary
flask

